+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
     - Release notes and changelog sections in manual (#1098)
     - ``setup_database.sh`` from SODAR Core
     - Enable ``appalerts`` app (#1124)
+    - Display relevant Django settings values in ``siteinfo`` app (#1123)
 - **Irodsbackend**
     - Support for ``IRODS_ROOT_PATH`` setting (#1067)
     - ``get_root_path()`` and ``get_projects_path()`` helpers (#1067)